Update doc and add maximum-query-records back
