+++ config/poli.properties
@@ -6,6 +6,9 @@
 # The maximum number of data source pool size
 # poli.datasource-maximum-pool-size=50
 
+# The maximum number of records returned in the JDBC result set. Default value is -1 which returns unlimited rows.
+# poli.maximum-query-records=-1
+
 # The display language.
 # poli.locale-language=en
