+++ src/main/java/com/shzlw/poli/util/CommonUtil.java
@@ -36,24 +36,4 @@ public final class CommonUtil {
     public static String getParamByAttrKey(String attrKey) {
         return "$user_attr[" + attrKey + "]";
     }
-
-    public static List<String> getQueryStatements(String sql) {
-        if (sql == null || sql.isEmpty()) {
-            return Collections.emptyList();
-        }
-
-        if (!sql.contains(";")) {
-            return Arrays.asList(sql);
-        }
-
-        List<String> statements = new ArrayList<>();
-        String[] sqlArray = sql.split(";");
-        for (String s : sqlArray) {
-            String t = s.trim();
-            if (!t.isEmpty()) {
-                statements.add(t + ";");
-            }
-        }
-        return statements;
-    }
 }
