+++ src/test/java/com/shzlw/poli/util/CommonUtilTest.java
@@ -1,5 +1,6 @@
 package com.shzlw.poli.util;
 
+import com.shzlw.poli.service.JdbcQueryServiceHelper;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -46,41 +47,5 @@ public class CommonUtilTest {
     public void testGetParamByAttrKey() {
         String rt = CommonUtil.getParamByAttrKey("attrKey");
         Assert.assertEquals("$user_attr[attrKey]", rt);
-    }
-
-    @Test
-    public void testGetQueryStatements_1() {
-        Assert.assertEquals(0, CommonUtil.getQueryStatements("").size());
-        Assert.assertEquals(0, CommonUtil.getQueryStatements(null).size());
-    }
-
-    @Test
-    public void testGetQueryStatements_2() {
-        List<String> sqls = CommonUtil.getQueryStatements("aaa");
-        Assert.assertEquals(1, sqls.size());
-        Assert.assertEquals(sqls.get(0), "aaa");
-    }
-
-    @Test
-    public void testGetQueryStatements_3() {
-        List<String> sqls = CommonUtil.getQueryStatements("aaa;");
-        Assert.assertEquals(1, sqls.size());
-        Assert.assertEquals(sqls.get(0), "aaa;");
-    }
-
-    @Test
-    public void testGetQueryStatements_4() {
-        List<String> sqls = CommonUtil.getQueryStatements("aaa;bbb;");
-        Assert.assertEquals(2, sqls.size());
-        Assert.assertEquals(sqls.get(0), "aaa;");
-        Assert.assertEquals(sqls.get(1), "bbb;");
-    }
-
-    @Test
-    public void testGetQueryStatements_5() {
-        List<String> sqls = CommonUtil.getQueryStatements("aaa;bbb;     ");
-        Assert.assertEquals(2, sqls.size());
-        Assert.assertEquals(sqls.get(0), "aaa;");
-        Assert.assertEquals(sqls.get(1), "bbb;");
     }
 }
